Move internal utils to a subpackage
