+++ core/src/commonMain/kotlin/com/episode6/mxo2/internal/MxoInstanceBuilder.kt
@@ -3,11 +3,13 @@ package com.episode6.mxo2.internal
 import com.episode6.mxo2.DependencyAlreadyMappedError
 import com.episode6.mxo2.MockspressoInstance
 import com.episode6.mxo2.api.Dependencies
+import com.episode6.mxo2.api.DynamicObjectMaker
 import com.episode6.mxo2.api.FallbackObjectMaker
 import com.episode6.mxo2.api.ObjectMaker
-import com.episode6.mxo2.api.DynamicObjectMaker
 import com.episode6.mxo2.defaultFallbackObjectMaker
 import com.episode6.mxo2.defaultRealObjectMaker
+import com.episode6.mxo2.internal.util.DependencyCacheBuilder
+import com.episode6.mxo2.internal.util.RealObjectRequestsList
 import com.episode6.mxo2.reflect.DependencyKey
 import com.episode6.mxo2.reflect.TypeToken
